+++ programs/clearing_house/src/controller/position.rs
@@ -664,7 +664,7 @@ pub fn update_settled_pnl(position: &mut PerpPosition, delta: i64) -> ClearingHo
 pub fn increase_open_bids_and_asks(
     position: &mut PerpPosition,
     direction: &PositionDirection,
-    base_asset_amount_unfilled: u128,
+    base_asset_amount_unfilled: u64,
 ) -> ClearingHouseResult {
     match direction {
         PositionDirection::Long => {
@@ -687,7 +687,7 @@ pub fn increase_open_bids_and_asks(
 pub fn decrease_open_bids_and_asks(
     position: &mut PerpPosition,
     direction: &PositionDirection,
-    base_asset_amount_unfilled: u128,
+    base_asset_amount_unfilled: u64,
 ) -> ClearingHouseResult {
     match direction {
         PositionDirection::Long => {
